final changes for reproducability of SAC
