+++ main.py
@@ -38,7 +38,7 @@ class RewardPrinter(BaseCallback):
                             self.logger.record(f"reward_components/{key}", value)
         return True
 
-def make_lift_env(dense=False):
+def make_lift_env():
     args = parse_args()
     env = RewardOverrideWrapper(
         robots="Panda",  
@@ -49,23 +49,37 @@ def make_lift_env(dense=False):
         reward_shaping=args.reward_shaping,
         control_freq=20,
         horizon=500,
-        dense=dense,
     )
     env = GymWrapper(env)
     env = Monitor(env)
     return env
 
+def make_door_env():
+    env = suite.make(
+        env_name="Door", 
+        robots="Panda",
+        has_renderer=False,
+        has_offscreen_renderer=False,
+        use_camera_obs=False,
+        use_object_obs=True,
+        reward_shaping=True,
+        control_freq=20,
+        horizon=500,    
+        reward_scale=1.0,
+    )
+    
+    gym_env = GymWrapper(env)
+    gym_env = Monitor(gym_env)
+    return gym_env
+
 def parse_args():
     parser = ArgumentParser()
-
+    parser.add_argument('--task', type=str, choices=['lift', 'door'], default='lift', help="Task to train on")
     parser.add_argument('--continue_train', action='store_true', help="Continue training from saved model")
-    parser.add_argument('--model', type=str, choices=['PPO', 'SAC'], default='PPO', help="RL algorithm to use")
-    parser.add_argument('--dense', action='store_true', help='Use dense reward')
-
     parser.add_argument('--reward_shaping', action='store_true', help="Enable reward shaping")
     parser.add_argument('--no_reward_shaping', dest='reward_shaping', action='store_false', help="Disable reward shaping")
     parser.set_defaults(reward_shaping=True)
-
+    parser.add_argument('--checkpoint', type=str, default="SAC_lift_dense_1mil.zip", help="Checkpoint file to load for training")
     parser.add_argument('--timesteps', type=int, default=500_000, help="Number of training timesteps")
 
     return parser.parse_args()
@@ -73,70 +87,46 @@ def parse_args():
 def main():
     args = parse_args()
     num_env = 1
-    vec_env = DummyVecEnv([lambda: make_lift_env(args.dense) for _ in range(num_env)])
+    if args.task == 'lift':
+        vec_env = DummyVecEnv([make_lift_env for _ in range(num_env)])
+    elif args.task == 'door':
+        vec_env = DummyVecEnv([make_door_env for _ in range(num_env)])
 
-    if args.model == 'PPO':
-        if args.continue_train:
-            print("Loading existing PPO model...")
-            model = PPO.load(
-                "PPO_dense.zip", 
-                env=vec_env,
-                verbose=1, 
-                tensorboard_log="./ppo_lift_tb/", 
-                seed=1,
-            )
-        else:
-            model = PPO(
-                "MlpPolicy", 
-                env=vec_env,
-                learning_rate=1e-4,
-                n_steps=2048,
-                batch_size=128,
-                ent_coef=0.01,
-                policy_kwargs={"net_arch": [256, 256]},
-                verbose=1, 
-                tensorboard_log="./ppo_lift_tb/", 
-                seed=1,
-            )
-    elif args.model == 'SAC':
-        if args.continue_train:
-            print("Loading existing SAC model...")
-            model = SAC.load(
-                "SAC_sparse.zip", 
-                env=vec_env,
-                verbose=1,
-                tensorboard_log="./sac_lift_tb/",
-                seed=1,
-            )
-        else:
-            model = SAC(
-                "MlpPolicy",
-                env=vec_env,
-                learning_rate=5e-4,
-                learning_starts=3300,
-                batch_size=128,
-                train_freq=2500,
-                gradient_steps=1000,
-                target_update_interval=5,
-                policy_kwargs={"net_arch": [256, 256]},
-                verbose=1,
-                tensorboard_log="./sac_lift_tb/",
-                seed=1,
-            )
+    if args.continue_train:
+        print("Loading existing SAC model...")
+        model = SAC.load(
+            args.checkpoint, 
+            env=vec_env,
+            verbose=1,
+            tensorboard_log=f"./sac_{args.task}_tb/",
+            seed=1,
+        )
+    else:
+        model = SAC(
+            "MlpPolicy",
+            env=vec_env,
+            learning_rate=5e-4,
+            learning_starts=3300,
+            batch_size=128,
+            train_freq=2500,
+            gradient_steps=1000,
+            target_update_interval=5,
+            policy_kwargs={"net_arch": [256, 256]},
+            verbose=1,
+            tensorboard_log=f"./sac_{args.task}_tb/",
+            seed=1,
+        )
 
     callback = RewardPrinter()
-    print(f"Training {args.model} on Lift environment for {args.timesteps} timesteps...")
-    print(f"Dense reward: {args.dense}, Reward shaping: {args.reward_shaping}")
+    print(f"Training SAC on {args.task} environment for {args.timesteps} timesteps...")
+    print(f"Reward shaping: {args.reward_shaping}")
 
     model.learn(total_timesteps=args.timesteps, callback=callback)
+    shaping = "dense" if args.reward_shaping else "sparse"
     if args.reward_shaping:
-        shaping = "shaping"
+        model.save(f'SAC_{args.task}_{shaping}_{args.timesteps}')
     else:
-        shaping = "no_shaping"
-    if args.dense:
-        model.save(f'{args.model}_dense_{shaping}_{args.timesteps}')
-    else:
-        model.save(f'{args.model}_sparse_{shaping}_{args.timesteps}')
+        model.save(f'SAC_{args.task}_{shaping}_{args.timesteps}')
 
     print("Testing trained model...")
     obs = vec_env.reset()
